SpaceDiceReloaded/pyproject.toml
[tool.poetry]
name = "spacedice-reloaded"
version = "2.0.0"
description = "Space-themed dice roller using NASA cosmic entropy"
authors = ["SpaceDice Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
gunicorn = "^21.0.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
numpy = "^2.0.0"
scipy = "^1.14.0"
statsmodels = "^0.14.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
